feat: add sqlite pooling and staggered jobs
